projects/otc_kg/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
neo4j>=5.14.0
mysql-connector-python>=8.2.0
pydantic>=2.5.0
python-dotenv>=1.0.0
